Configure X-Ray for ETL Lambda Functions (#157)
After enabling X-Ray on the State Machine, the next step is to enable the X-Ray daemon so the Lambda functions can send data

- Enable Tracing for Lambda Functions
- Set JSON Logging type for the Lambda Cloudwatch logs
- Fix the ResultWriter Parameter for the Prefix
  - Because we're not using any substitutions `Prefix.$` results in an error 
- Add requestId to logging dict via the decorator
- Add POWERTOOLS_SERVICE_NAME env var to Lambdas
- Add Description text to lambdas to make them easier to track
- Add `@tracer.capture_lambda_handler` decorator to all ETL Lambda
TODO: It seems that the Map's Lambda Functions are not reporting any X-Ray Traces. 
This seems to be because it's a Distributed Map. 
We need to figure out how to supply the Trace ID from the parent into the map and init the X Ray SDK with it. 
We may need to use the X-Ray SDK Directly instead of via the Powertool