fix(txc-parser): implement DaysOfWeek parsing bug to match current bods implementation
Currently BODs assumes that a RegularDayType that has `<DaysOfWeek />` means all days are false.

This contradicts the TXC 2.1 spec which says that it means MondayToSunday

However the currently implementation and the BODs timetable generator tool assumes that it means all days are false

This PR implements the incorrect behavior to match the current setup.

There is a test marked as fail for the correct behavior and and there's a var called `incorrect_implementation = True` to ease the effort in implementing the correct behavior in the future

JIRA: https://kpmgengineering.atlassian.net/browse/BODS-8037