feat(PostSchemaCheck): Check service code exists in a published dataset (#263)
This is a new post schema rule added to stop services that are already
published.

Key Details:

Add a new function to read the `TXCData.Services` which is the list of
service codes present in the current dataset being published. Check if
any or all the services are present in `OrganisationTxcFileAttributes`
table, get all the revision IDs corresponding to the service codes and
check if the revision is live and the dataset is published. If this
check returns a dataset then return a list of `ValidationResult` with
the dataset ID and all the service codes present in the dataset.

JIRA: https://kpmgengineering.atlassian.net/browse/BODS-399

---------

Co-authored-by: andyecrow <38900570+andyecrow@users.noreply.github.com>