refactor(archiver): Use SqlDB interface replace old DbManager and remove unused code (#260)
Use latest SqlDB interface and refactor the archiver module, Remove
unused code and update the test. This lead to refactor the periodic task
lambdas and its test case

Key Details:
- upgrade `src/boilerplate/common_layer/archiver.py` using `SqlDB` class
- Removed following package which is no longer needed
   - `src/boilerplate/common_layer/db/repositories`
   - `tests/boilerplate/common_layer/db/repositories`
- Updated unit test for archiver
- Refactored periodic task lambdas
- Updated unit test case for periodic task lambdas, remove duplicate
tests
- Removed BodsDB and DbManager class, module and its tests

TODO: GTFS_API_ACTIVE flag is not populated to serverless from django, I
learnt that this flag may be removed. Need to check with amit on this
flag. it is for only gtfstr data.