fix(PipelineErrorCode): Create error code if does not exists in PipelineErrorCode table (#318)
PipelineErrorCode model is populated with a new record for the error
codes being raised in the lambdas. In Django, the column in the model is
defined as an enum of choices but since data is not inserted into the
FileProcessingResult table through Django, it makes sense to just add it
here.


JIRA: https://kpmgengineering.atlassian.net/browse/BODS-8332