fix(PublishDataset): BODS-8333 link published revision to a dataset (#323)
This change is to update `live_revision_id` on the dataset to point to
newly published revision. Although, a revision is marked as `live` after
new data is found, it's not linked to the dataset. This causing the
update job to always find difference in the files on BODS and on the
URL.


JIRA: https://kpmgengineering.atlassian.net/browse/BODS-8333