fix(timetables-etl): Match Celery Wait Time Calculation (#335)
The celery implementation uses the following logic:

1. Add WaitTime either from the From or the To depending on where it is
present
If both From and To is present, pick the <To> WaitTime
2. The first stop will always take the <From> WaitTime (there cannot be
a <To> WaitTime for the first stop)
3. The last stop will not take the <To> WaitTime (there cannot be a
<From> WaitTime for the last stop). The reason we don't add WaitTime to
the last stop is because we want to capture arrival time at the final
stop. The bus can wait hours at the last stop but we don't need to show
this on the timetable.
5. If one waittime is <WaitTime>PT0S</WaitTime> we consider it not
present


JIRA: https://kpmgengineering.atlassian.net/browse/BODS-8398