fix(timetables-etl): BODS-8417 further filter out dates based on day of the week (#350)
- Operating date exception table should only be populated if any special
days of operation period falls on a day the vehicle is not already
running.

- Non Operating date exception table should only be populated if any
special days of non operation period falls on a day the vehicle is
running.

- Operating date exception table should only be populated if any Bank
Holiday operation falls on a day the vehicle is not already running.

- Non Operating date exception table should only be populated if any
Bank Holiday non operation period falls on a day the vehicle is running.

Refer these two JIRA tickets where the logic is defined in detail
https://kpmgengineering.atlassian.net/browse/BODS-2287 and
https://kpmgengineering.atlassian.net/browse/BODS-2286

JIRA: https://kpmgengineering.atlassian.net/browse/BODS-8417