fix(ExceptionHandler): Handle unexpected error formats (#419)
Right now platform errors such as timeouts or unexpected runtime.exit
errors caused OOM exceptions aren't handled well. In the former, the
state machine definition fails to parse the cause to json, preventing
the exception handler from running.
In the latter case, we get validation errors when parsing the event to
the `ExceptionHandlerInputData` model.

In this PR we move the responsibility of parsing the error cause to json
to the application code, and handle the case that it doesn't match the
expected format.

---------

Co-authored-by: Andy Crow <andy.crow@doza.consulting>
Co-authored-by: Jonathan Nakandala <Jonathan.Nakandala@kpmg.co.uk>