fix(PTIValidation): Optimize line validation to avoid timeouts (#538)
PTI lambda times out for a large xml file with lots of journeys in it.
The likely cause of this is repeated XPath queries on the XML root to
get each journey pattern section and then the stops of that journey.

So instead of querying per journey pattern ref, I am going to try to
fetch all relevant data in two of wider XPath queries and then build
mapping in memory.

Key Details:

- Create two dictionaries in the base class of `LinesValidator`
- Store journey patterns and journey pattern sections in a dict and get
the required data out of it using the line ref when needed.

JIRA: https://kpmgengineering.atlassian.net/browse/BODS-9138

---------

Co-authored-by: Andy Crow <andy.crow@doza.consulting>